Add HeartRate endpoint and JSON unmarshalling
This update expands the 'heartrates.go' by adding methods for JSON unmarshalling for HeartRate items and HeartRates. It adds a GetHeartRates function to the Client. An accompanying example code for accessing heart rate data is also included in a new 'heartrates_example.go' file. The HeartRate endpoint URL is added in 'config.go'.