Add SpO2 readings retrieval functionality
A new file, 'spo2.go', has been added to handle the retrieval of SpO2 readings from the specified API endpoint. This includes the setup of relevant data structures and unmarshalling procedures, as well as the method 'GetSpo2Readings' that retrieves the readings for a given date range. Furthermore, the 'Spo2Url' has been added to 'config.go' to support this new functionality.