Update GetHeartRates to handle next_token
This update modifies the GetHeartRates function to allow the use of the "next_token" parameter for paginated results. Specifically, the heartrates_example.go and heartrates_test.go files have been changed to handle this parameter. The "next_token" is now also handled in the testing scenario in the heartrates_test.go file. This change provides more flexibility and functionality for clients fetching heart rate data.