Add functionality and examples for fetching tags
A new `Tags` struct that relates to the `enhanced_tag` endpoint in the Oura API has been added. This includes JSON marshalling/unmarshalling methods as well as methods to fetch individual and multiple tag records. The `GetTags` and `GetTag` methods have been added to the `Client` for easier access. Necessary examples on how to use this functionality have been provided in `tag_example.go`.