Update Tag struct and add tag tests
Modified the Tag struct in tags.go to make StartTime, EndTime, StartDay, and EndDay pointers. This allows these fields to be nil, improving the robustness of the application especially when dealing with incomplete data. Accompanying these changes, tests were also added for the Tag struct in a new file tag_test.go to ensure it is working correctly.